requirements.txt
anyascii==0.3.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.7
boto3==1.21.32
botocore==1.24.32
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
contractions==0.1.68
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
idna==3.3
importlib-resources==5.6.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
kafka-python3==3.0.0
kazoo==2.5.0
kiwisolver==1.4.0
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyahocorasick==1.4.4
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
tabulate==0.8.9
terminado==0.13.3
testpath==0.6.0
textsearch==0.0.21
thinc==8.0.15
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
tweepy==4.8.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
wordcloud==1.8.1
zipp==3.7.0

pykafka~=2.8.0